[SPARK-39966][SQL] Use V2 Filter in SupportsDelete
### What changes were proposed in this pull request?
Migrate SupportsDelete to use V2 Filter

### Why are the changes needed?
this is part of the V2Filter migration work

### Does this PR introduce _any_ user-facing change?
Yes
add `SupportsDeleteV2`

### How was this patch tested?
new tests

Closes #37393 from huaxingao/v2delete.

Authored-by: huaxingao <huaxin_gao@apple.com>
