+++ sql/catalyst/src/main/java/org/apache/spark/sql/connector/catalog/SupportsDeleteV2.java
@@ -18,19 +18,17 @@
 package org.apache.spark.sql.connector.catalog;
 
 import org.apache.spark.annotation.Evolving;
+import org.apache.spark.sql.connector.expressions.filter.AlwaysTrue;
 import org.apache.spark.sql.connector.expressions.filter.Predicate;
-import org.apache.spark.sql.internal.connector.PredicateUtils;
-import org.apache.spark.sql.sources.AlwaysTrue;
-import org.apache.spark.sql.sources.Filter;
 
 /**
  * A mix-in interface for {@link Table} delete support. Data sources can implement this
  * interface to provide the ability to delete data from tables that matches filter expressions.
  *
- * @since 3.0.0
+ * @since 3.4.0
  */
 @Evolving
-public interface SupportsDelete extends SupportsDeleteV2 {
+public interface SupportsDeleteV2 extends TruncatableTable {
 
   /**
    * Checks whether it is possible to delete data from a data source table that matches filter
@@ -39,24 +37,25 @@ public interface SupportsDelete extends SupportsDeleteV2 {
    * Rows should be deleted from the data source iff all of the filter expressions match.
    * That is, the expressions must be interpreted as a set of filters that are ANDed together.
    * <p>
-   * Spark will call this method at planning time to check whether {@link #deleteWhere(Filter[])}
+   * Spark will call this method at planning time to check whether {@link #deleteWhere(Predicate[])}
    * would reject the delete operation because it requires significant effort. If this method
-   * returns false, Spark will not call {@link #deleteWhere(Filter[])} and will try to rewrite
+   * returns false, Spark will not call {@link #deleteWhere(Predicate[])} and will try to rewrite
    * the delete operation and produce row-level changes if the data source table supports deleting
    * individual records.
    *
-   * @param filters filter expressions, used to select rows to delete when all expressions match
+   * @param predicates V2 filter expressions, used to select rows to delete when all expressions
+   *                  match
    * @return true if the delete operation can be performed
    *
-   * @since 3.1.0
+   * @since 3.4.0
    */
-  default boolean canDeleteWhere(Filter[] filters) {
+  default boolean canDeleteWhere(Predicate[] predicates) {
     return true;
   }
 
   /**
    * Delete data from a data source table that matches filter expressions. Note that this method
-   * will be invoked only if {@link #canDeleteWhere(Filter[])} returns true.
+   * will be invoked only if {@link #canDeleteWhere(Predicate[])} returns true.
    * <p>
    * Rows are deleted from the data source iff all of the filter expressions match. That is, the
    * expressions must be interpreted as a set of filters that are ANDed together.
@@ -67,27 +66,18 @@ public interface SupportsDelete extends SupportsDeleteV2 {
    * To reject a delete implementations should throw {@link IllegalArgumentException} with a clear
    * error message that identifies which expression was rejected.
    *
-   * @param filters filter expressions, used to select rows to delete when all expressions match
+   * @param predicates predicate expressions, used to select rows to delete when all expressions
+   *                  match
    * @throws IllegalArgumentException If the delete is rejected due to required effort
    */
-  void deleteWhere(Filter[] filters);
-
-  default boolean canDeleteWhere(Predicate[] predicates) {
-    Filter[] v1Filters = PredicateUtils.toV1(predicates);
-    if (v1Filters.length < predicates.length) return false;
-    return this.canDeleteWhere(v1Filters);
-  }
-
-  default void deleteWhere(Predicate[] predicates) {
-    this.deleteWhere(PredicateUtils.toV1(predicates));
-  }
+  void deleteWhere(Predicate[] predicates);
 
   @Override
   default boolean truncateTable() {
-    Filter[] filters = new Filter[] { new AlwaysTrue() };
-    boolean canDelete = canDeleteWhere(filters);
+    Predicate[] predicates = new Predicate[] { new AlwaysTrue() };
+    boolean canDelete = canDeleteWhere(predicates);
     if (canDelete) {
-      deleteWhere(filters);
+      deleteWhere(predicates);
     }
     return canDelete;
   }
