+++ sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/v2/DeleteFromTableExec.scala
@@ -19,12 +19,12 @@ package org.apache.spark.sql.execution.datasources.v2
 
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions.Attribute
-import org.apache.spark.sql.connector.catalog.SupportsDelete
-import org.apache.spark.sql.sources.Filter
+import org.apache.spark.sql.connector.catalog.SupportsDeleteV2
+import org.apache.spark.sql.connector.expressions.filter.Predicate
 
 case class DeleteFromTableExec(
-    table: SupportsDelete,
-    condition: Array[Filter],
+    table: SupportsDeleteV2,
+    condition: Array[Predicate],
     refreshCache: () => Unit) extends LeafV2CommandExec {
 
   override protected def run(): Seq[InternalRow] = {
